clang-tidy

+++ src/multi/content_identifier_codec.cpp
@@ -97,8 +97,8 @@ namespace libp2p::multi {
     std::vector<uint8_t> bytes;
     // Reserve space for CID version size + content-type size + multihash size
     bytes.reserve(1 + 1 + mhash.toBuffer().size());
-    bytes.push_back(1);  // CID version
-    bytes.push_back(static_cast<uint8_t>(content_type)); // Content-Type
+    bytes.push_back(1);                                   // CID version
+    bytes.push_back(static_cast<uint8_t>(content_type));  // Content-Type
     std::copy(mhash.toBuffer().begin(), mhash.toBuffer().end(),
               std::back_inserter(bytes));  // multihash data
     return bytes;
@@ -118,12 +118,13 @@ namespace libp2p::multi {
     auto version = version_opt.value().toUInt64();
     if (version == 1) {
       auto version_length = UVarint::calculateSize(bytes);
-      auto multicodec_opt = UVarint::create(bytes.subspan(version_length));
+      auto multicodec_opt = UVarint::create(
+          bytes.subspan(static_cast<ptrdiff_t>(version_length)));
       if (!multicodec_opt) {
         return DecodeError::EMPTY_MULTICODEC;
       }
-      auto multicodec_length =
-          UVarint::calculateSize(bytes.subspan(version_length));
+      auto multicodec_length = UVarint::calculateSize(
+          bytes.subspan(static_cast<ptrdiff_t>(version_length)));
       OUTCOME_TRY(hash,
                   Multihash::createFromBytes(
                       bytes.subspan(version_length + multicodec_length)));
