+++ src/protocol/kademlia/impl/content_id.cpp
@@ -19,7 +19,8 @@ namespace libp2p::protocol::kademlia {
   ContentId::ContentId(std::string_view str) {
     auto digest_res = crypto::sha256(gsl::make_span(
         // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
-        reinterpret_cast<const uint8_t *>(str.data()), str.size()));
+        reinterpret_cast<const uint8_t *>(str.data()),
+        static_cast<ptrdiff_t>(str.size())));
     BOOST_ASSERT(digest_res.has_value());
 
     auto mhash_res = libp2p::multi::Multihash::create(
@@ -36,7 +37,8 @@ namespace libp2p::protocol::kademlia {
   boost::optional<ContentId> ContentId::fromWire(std::string_view str) {
     gsl::span<const uint8_t> bytes(
         // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
-        reinterpret_cast<const uint8_t *>(str.data()), str.size());
+        reinterpret_cast<const uint8_t *>(str.data()),
+        static_cast<ptrdiff_t>(str.size()));
     return fromWire(bytes);
   }
